requirements.txt
# 合同审核 Skill 依赖

# 文档解析
pdfplumber>=0.10.0      # PDF 解析
python-docx>=1.0.0      # Word 文档解析

# HTTP 客户端 (LLM API 调用)
httpx>=0.25.0

# 可选: 如果使用 OpenAI 兼容接口
# openai>=1.0.0
